ingrain_model_server/pyproject.toml
[project]
name = "ingrain-model-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2025.7.9",
    "fastapi==0.116.1",
    "hf-transfer==0.1.9",
    "onnx==1.18.0",
    "onnxconverter-common>=1.13.0",
    "onnxoptimizer>=0.3.13",
    "onnxscript>=0.3.2",
    "open-clip-torch==2.32.0",
    "orjson==3.10.18",
    "pillow==11.3.0",
    "pycurl==7.45.6",
    "pydantic==2.11.7",
    "sentence-transformers==5.0.0",
    "timm==1.0.17",
    "tokenizers==0.21.2",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "transformers==4.53.2",
    "tritonclient[grpc]>=2.58.0",
    "uvicorn==0.35.0",
    "validators==0.35.0",
]
